+++ internal/manifest/fixtures/maven/not-pom.txt
@@ -1,0 +1,1 @@
+this is not a pom.xml file!
